+++ src/api/controllers/stats/hourly/hourlyStats.spec.js
@@ -7,6 +7,7 @@ import mockery from "mockery";
 
 import { FakeAuthMiddleware } from '../../../test-helper';
 import HourlyStats from './';
+import contentNegotiation from '../../../middleware/content-negotiation-middleware';
 
 import { HourlyStatsProvider } from '../../../../data/mongodb';
 
@@ -30,6 +31,8 @@ describe('HourlyStats API endpoints', () => {
     mockery.registerMock('../../common/logger', { log: console.log });
 
     app = express();
+    app.use(contentNegotiation());
+    
     request = supertest(app);
     const db = {
       collection: () => {
@@ -70,4 +73,35 @@ describe('HourlyStats API endpoints', () => {
         }
       });
   });
+
+  it('should get hourly consume for a given deviceId', (done) => {
+    const date = new Date();
+
+    const stub = sinon.stub(hourlyStatsProvider, 'getHourlyStatByDevice')
+      .returns(Promise.resolve(
+        [{
+          _id: 123,
+          deviceId: '11:22:33:44:55:66',
+          power: 20,
+        }]),
+    );
+
+    HourlyStats(app, [FakeAuthMiddleware(['gwtest'])()], { hourlyStatsProvider });
+
+    request
+      .get(`/api/stats/hourly/11:22:33:44:55:66?gw=gwtest&date=${date}`)
+      .expect(200, (err, res) => {
+        if (err) {
+          done(err);
+        } else {
+          stub.calledWith(sinon.match.any, ['gwtest'], '11:22:33:44:55:66')
+            .should.be.true;
+          const response = res.body;
+          response.length.should.be.eq(1);
+          response[0].deviceId.should.be.eq('11:22:33:44:55:66');
+          response[0].power.should.be.eq(20);
+          done();
+        }
+      });
+  })
 });