+++ src/api/controllers/stats/hourly/index.js
@@ -1,13 +1,12 @@
 import express from 'express';
-import nodeExcel from 'excel-export';
 import * as _ from 'lodash';
 import logger from '../../../../common/logger';
-import { getDate, getOverlapped, getGroupField, createExcelConf } from '../../../api-utils';
+import { getDate, getOverlapped, getGroupField } from '../../../api-utils';
 import { transformHourlyStat } from './hourlyStatTransformer';
 
-function getHourlyDates(date, startHour) {
+function getHourlyDates(date) {
   const todayHours = [];
-  date.setUTCHours(startHour || 7);
+  date.setUTCHours(0);
   while (date.getUTCHours() < 23) {
     date.setUTCHours(date.getUTCHours() + 1);
     todayHours.push(new Date(date));
@@ -62,35 +61,45 @@ export default function (app, middlewares, { hourlyStatsProvider }) {
     const date = getDate(req);
     const ownedGws = req.user.app_metadata.gateways;
     const reqGateways = req.query.gw;
-    const format = req.query.format;
     const group = getGroupField(req);
-    const startHour = req.query.startHour;
 
     const gws = getOverlapped(ownedGws, reqGateways);
 
     hourlyStatsProvider
-      .getHourlyStat(getHourlyDates(date, startHour), gws, group)
+      .getHourlyStat(getHourlyDates(date), gws, group)
       .then((stat) => {
         if (stat.length === 0) {
           res.sendStatus(204);
         } else {
-          const result = stat.filter(s => !_.isEmpty(s))
-          .map(s => transformHourlyStat(s));
-          if (format === 'excel') {
-            const conf = createExcelConf(_.orderBy(result, 'hour'));
-            const excel = nodeExcel.execute(conf);
-            res.setHeader('Content-Type', 'application/vnd.openxmlformats');
-            res.setHeader('Content-Disposition', 'attachment');
-            res.end(excel, 'binary');
-          } else {
-            res.json(result);
-          }
+          const result = stat.filter(s => !_.isEmpty(s));
+          res.sendData(result, {
+            'application/json': d => d.map(s => transformHourlyStat(s)),
+            'application/vnd.ms-excel': d => d.map(s => transformHourlyStat(s)),
+          });
         }
       })
       .catch((err) => {
         logger.log('error', err);
         res.sendStatus(500);
       });
+  });
+
+  router.get('/:deviceId', middlewares, (req, res) => {
+    const deviceId = req.params.deviceId;
+    const date = req.query.date;
+    const reqGateways = req.query.gw;
+    const ownedGws = req.user.app_metadata.gateways;
+
+    const gws = getOverlapped(ownedGws, reqGateways);
+
+    hourlyStatsProvider.getHourlyStatByDevice(date, gws, deviceId)
+    .then((stat) => {
+      if (stat.length === 0) {
+        res.sendStatus(204);
+      } else {
+        res.json(stat.map(s => transformHourlyStat(s)));
+      }
+    });
   });
 
   return router;