+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-<a name="2.1.2"></a>
-## 2.1.2 (2017-09-07)
+<a name="2.2.0"></a>
+# 2.2.0 (2017-09-13)
 
 
 ### Bug Fixes
@@ -27,11 +27,13 @@
 
 * Add alert builder to manage alert objects creation ([2f591d2](https://gitlab.com/smart-office-iot/iot-project-server/commit/2f591d2))
 * Add alert on power status change failure ([1f361a7](https://gitlab.com/smart-office-iot/iot-project-server/commit/1f361a7))
+* Add content negotiation middleware to serialize API response based on Accept header value ([a73fea0](https://gitlab.com/smart-office-iot/iot-project-server/commit/a73fea0))
 * Add createIndex to DataProvider to encapsulate index creation ([b38cffe](https://gitlab.com/smart-office-iot/iot-project-server/commit/b38cffe))
 * Add delete alert endpoint ([9ff9228](https://gitlab.com/smart-office-iot/iot-project-server/commit/9ff9228))
 * Add device commands endpoint ([297aa8c](https://gitlab.com/smart-office-iot/iot-project-server/commit/297aa8c))
 * Add endpoint to get single device data ([e416b35](https://gitlab.com/smart-office-iot/iot-project-server/commit/e416b35))
 * Add endpoint to mark alert read/unread ([8331828](https://gitlab.com/smart-office-iot/iot-project-server/commit/8331828))
+* Add endpoint to retrieve hour consume for single device ([7ae94e6](https://gitlab.com/smart-office-iot/iot-project-server/commit/7ae94e6))
 * Add energy alerts (power == 0 when device is on) aggregation for same device/gateway in given ([7c4e7f0](https://gitlab.com/smart-office-iot/iot-project-server/commit/7c4e7f0))
 * Add gateway auth logic (based on authToken field in Gateways collection) ([0252395](https://gitlab.com/smart-office-iot/iot-project-server/commit/0252395))
 * Add gateway field to device DTO ([32a2397](https://gitlab.com/smart-office-iot/iot-project-server/commit/32a2397))